Reorganize sanity loaders
