+++ sanity/loader/loadQuery.ts
@@ -1,10 +1,10 @@
 import 'server-only'
 
 import { draftMode } from 'next/headers'
-import * as queryStore from '@sanity/react-loader'
 import type { ContentSourceMap, QueryOptions, QueryParams, SanityClient, SanityDocument } from "@sanity/client";
 import { client } from "@/sanity/lib/client";
 import { token } from '@/sanity/lib/token';
+import { queryStore } from '@/sanity/loader/queryStore';
 
 export const serverClient: SanityClient = client.withConfig({
 	token,
